Version 0.1.0
- The DTO for the following packages are now available: kali-tools-802-11, kali-tools-bluetooth, kali-tools-crypto-stego.

- All the DTO for kali-tools are in /lib/kali_tools/dto

- Every DTO have: enum to identify each package, a String version from Map to have the proper String name, an extension to get the String from enum and a method to show all packages from selected group to install.

- kali_tools/select_tools_to_install.dart

The method selectToolsToInstall show all packages from given enum and get all packages selected from user. It has also the option to show help, install all, remove all and go back in the menu.

- print_title.dart
It prints the title of the project.
ESSENTIAL

- setup/pkg/install_flatpak.dart
 Added flatpak repo

Package version updated to 0.1.0

Download run.sh and run it as sudo to try it out