Added documentation and versioning packages
- All project files have documentation

- Removed Flatpak from the project

- Fixed some DTO tools for incorrect name

- Created kali_tools/services/package_manager.dart
Created PackageVersion to track version and install for each package (APT and Kali)
Created getPackageAvailability to get version for APT and Kali of the package. It also get if the package is already installed

- services/select_tools_to_install.dart
Removed ToolSelectionResultType
Now each package show the version for APT and Kali

- setup/pkg/check_installed_pkg.dart
Created checkAllPackageVersions(). This function get the policy info for all given packages and then set packageAvailability[package] with getPackageAvailability(package).
packageAvailability is a Map that stores versioning for all packages